Slot functionality

+++ src/contexts/AppContext.jsx
@@ -313,6 +313,7 @@ const AppContextProvider = ({ children }) => {
     morningHour: "",
     morningMinute: "",
     morningAmpm: "AM",
+    slots: "",
   });
 
   //to save morning time form input
@@ -331,6 +332,7 @@ const AppContextProvider = ({ children }) => {
     noonHour: "",
     noonMinute: "",
     noonAmpm: "PM",
+    slots: "",
   });
 
   //to save noon time form input
@@ -347,6 +349,7 @@ const AppContextProvider = ({ children }) => {
   //function to save morning booking time doc on sign up
   //function to save morning booking time doc on sign up
   //function to save morning booking time doc on sign up
+  const [activeRideChange, setActiveRideChange] = useState(false);
   const [updatedTime, setUpdatedTime] = useState(false);
 
   //to save booking time from db
@@ -380,10 +383,10 @@ const AppContextProvider = ({ children }) => {
       }
     };
     getMorningBookingTime();
-  }, [updatedTime]);
+  }, [updatedTime, activeRideChange]);
 
   //to send created notes to db
-  const createMorningBookingTimeDocument = async (time, createdAt) => {
+  const createMorningBookingTimeDocument = async (time, createdAt, slots) => {
     setLoader(true);
 
     try {
@@ -410,6 +413,7 @@ const AppContextProvider = ({ children }) => {
             morningBookingTimesFromDb.length + 1
           }`,
           createdAt: createdAt,
+          slots: slots,
         }
       );
       console.log("morning booking time created");
@@ -429,7 +433,11 @@ const AppContextProvider = ({ children }) => {
     setLoader(true);
 
     try {
-      await createMorningBookingTimeDocument(morningTime, formattedDate);
+      await createMorningBookingTimeDocument(
+        morningTime,
+        formattedDate,
+        morningForm.slots
+      );
       setLoader(false);
     } catch (error) {
       setLoader(false);
@@ -471,10 +479,10 @@ const AppContextProvider = ({ children }) => {
       }
     };
     getNoonBookingTime();
-  }, [updatedTime]);
+  }, [updatedTime, activeRideChange]);
 
   //to send created notes to db
-  const createNoonBookingTimeDocument = async (time, createdAt) => {
+  const createNoonBookingTimeDocument = async (time, createdAt, slots) => {
     setLoader(true);
 
     try {
@@ -499,6 +507,7 @@ const AppContextProvider = ({ children }) => {
             noonBookingTimesFromDb.length + 1
           }`,
           createdAt: createdAt,
+          slots: slots,
         }
       );
       console.log("noon booking time created");
@@ -518,7 +527,11 @@ const AppContextProvider = ({ children }) => {
     setLoader(true);
 
     try {
-      await createNoonBookingTimeDocument(noonTime, formattedDate);
+      await createNoonBookingTimeDocument(
+        noonTime,
+        formattedDate,
+        noonForm.slots
+      );
       await getNoonBookingTime();
       added = false;
       setLoader(false);
@@ -541,8 +554,16 @@ const AppContextProvider = ({ children }) => {
 
   //to delete morning time
   function handleDeleteMorningTime(id) {
-    deleteMorningTimeDoc(id);
-    setUpdatedTime((prev) => !prev);
+    let mark = "Delete morning time?";
+
+    if (confirm(mark) == true) {
+      deleteMorningTimeDoc(id);
+      setUpdatedTime((prev) => !prev);
+      if (morningBookingTimesFromDb.length === 1) {
+        localStorage.removeItem("morningTimes");
+        window.location.reload();
+      }
+    }
   }
 
   //to delete time from db
@@ -557,8 +578,16 @@ const AppContextProvider = ({ children }) => {
 
   //to delete time
   function handleDeleteNoonTime(id) {
-    deleteNoonTimeDoc(id);
-    setUpdatedTime((prev) => !prev);
+    let mark = "Delete afternoon time?";
+
+    if (confirm(mark) == true) {
+      deleteNoonTimeDoc(id);
+      setUpdatedTime((prev) => !prev);
+      if (noonBookingTimesFromDb.length === 1) {
+        localStorage.removeItem("noonTimes");
+        window.location.reload();
+      }
+    }
   }
 
   //to get number of users
@@ -582,6 +611,56 @@ const AppContextProvider = ({ children }) => {
     }
     getUsers();
   }, []);
+
+  //to get total number of rides
+  const [allRides, setAllRides] = useState([]);
+  useEffect(() => {
+    async function getAllRides() {
+      setLoader(true);
+
+      try {
+        const querySnapshot = await getDocs(collection(db, "rideHistory"));
+        let rides = [];
+        querySnapshot.forEach((doc) => {
+          rides.push(doc.data());
+        });
+        setAllRides(rides);
+      } catch (err) {
+        console.log(err);
+      } finally {
+        setLoader(false);
+      }
+    }
+    getAllRides();
+  }, []);
+
+  //to get number of rides today
+  const [ridesToday, setridesToday] = useState([]);
+
+  useEffect(() => {
+    if (user) {
+      const getRidesToday = async () => {
+        setLoader(true);
+        const userQuery = query(
+          collection(db, "rideHistory"),
+          where("createdAt", "==", `${formattedDate}`)
+        );
+        try {
+          const querySnapshot = await getDocs(userQuery);
+          let ride = [];
+          querySnapshot.forEach((doc) => {
+            ride.push(doc.data());
+          });
+          setridesToday(ride);
+        } catch (err) {
+          console.log(err.message);
+        } finally {
+          setLoader(false);
+        }
+      };
+      getRidesToday();
+    }
+  }, [currentUserFromDb]);
 
   //to save nprice form input
   const [priceForm, setPriceForm] = useState({
@@ -772,15 +851,13 @@ const AppContextProvider = ({ children }) => {
     }
   };
 
-  const [activeRideChange, setActiveRideChange] = useState(false);
-
   //to get and store active rides
   const [activeRidesFromDb, setActiveRidesFromDb] = useState(
     JSON.parse(localStorage.getItem("activeRide")) || []
   );
   useEffect(() => {
     setActiveRidesFromDb(JSON.parse(localStorage.getItem("activeRide")));
-  }, [currentUserFromDb]);
+  }, [currentUserFromDb, activeRideChange]);
 
   //to get user's active ride saved in db
   useEffect(() => {
@@ -800,7 +877,6 @@ const AppContextProvider = ({ children }) => {
           ride.length > 0 &&
             localStorage.setItem("activeRide", JSON.stringify(ride));
           ride.length > 0 && setActiveRidesFromDb(ride);
-          // setActiveRideChange((prev) => !prev);
         } catch (err) {
           console.log(err.message);
         } finally {
@@ -810,7 +886,7 @@ const AppContextProvider = ({ children }) => {
       };
       getActiveRides();
     }
-  }, [currentUserFromDb]);
+  }, [currentUserFromDb, activeRideChange]);
 
   //to get and store ride history
   const [rideHistoryFromDb, setRideHistoryFromDb] = useState(
@@ -818,7 +894,7 @@ const AppContextProvider = ({ children }) => {
   );
   useEffect(() => {
     setRideHistoryFromDb(JSON.parse(localStorage.getItem("rideHistory")));
-  }, [currentUserFromDb]);
+  }, [currentUserFromDb, activeRideChange]);
 
   //to get user's active ride saved in db
   useEffect(() => {
@@ -838,7 +914,6 @@ const AppContextProvider = ({ children }) => {
           ride.length > 0 &&
             localStorage.setItem("rideHistory", JSON.stringify(ride));
           ride.length > 0 && setRideHistoryFromDb(ride);
-          // setActiveRideChange((prev) => !prev);
         } catch (err) {
           console.log(err.message);
         } finally {
@@ -847,11 +922,12 @@ const AppContextProvider = ({ children }) => {
       };
       getRidesHistory();
     }
-  }, [currentUserFromDb]);
+  }, [currentUserFromDb, activeRideChange]);
 
   const [bookingSuccess, setBookingSuccess] = useState(false);
   function cloaseSuccessModal() {
     setBookingSuccess(false);
+    setActiveRideChange((prev) => !prev);
   }
 
   const clearActiveRideDoc = async (id) => {
@@ -863,16 +939,25 @@ const AppContextProvider = ({ children }) => {
     }
   };
 
-  function markCompleted() {
-    let mark = "Mark ride as completed?";
+  async function markCompleted() {
+    setLoader(true);
 
-    if (confirm(mark) == true) {
-      localStorage.removeItem("activeRide");
-      clearActiveRideDoc(activeRidesFromDb[0].id);
-      window.location.reload();
-      setActiveRideChange((prev) => !prev);
+    try {
+      let mark = "Mark ride as completed?";
+
+      if (confirm(mark) == true) {
+        localStorage.removeItem("activeRide");
+        await clearActiveRideDoc(activeRidesFromDb[0].id);
+        window.location.reload();
+        setActiveRideChange((prev) => !prev);
+      }
+    } catch (err) {
+      console.log(err);
+    } finally {
+      setLoader(false);
     }
   }
+  // console.log(activeRidesFromDb[0].id);
 
   return (
     <AppContext.Provider
@@ -880,6 +965,7 @@ const AppContextProvider = ({ children }) => {
         regForm,
         handleRegChange,
         loader,
+        setLoader,
         register,
         showPassword,
         togglePassword,
@@ -923,6 +1009,8 @@ const AppContextProvider = ({ children }) => {
         activeRidesFromDb,
         markCompleted,
         rideHistoryFromDb,
+        allRides,
+        ridesToday,
       }}
     >
       {children}
